WIP more progress on ORM, but... locking issues?

+++ helios/types.go
@@ -5,28 +5,28 @@ import (
 )
 
 type Event struct {
-  EventId           int
+  EventId           int64
   Name              string
   ExpectedFrequency string
 }
 
 type Threshold struct {
-  ThresholdId       int
-  EventId           int
+  ThresholdId       int64
+  EventId           int64
   Name              string
   SeverityLimit     float32
   SeverityFunction  string
 }
 
 type EventOccurrence struct {
-  EventOccurrenceId int
-  EventId           int
+  EventOccurrenceId int64
+  EventId           int64
   OccurredAt        time.Time
 }
 
 type ThresholdCheck struct {
-  ThresholdCheckId  int
-  ThresholdId       int
+  ThresholdCheckId  int64
+  ThresholdId       int64
   CheckedAt         time.Time
   Severity          float32
 }
